added better way of sending response usign ResponseEntity

+++ EmployeeApp/Test-Jpa/src/main/java/com/corona/TestJpa/service/EmployeeServiceImplementation/EmployeeServiceImpl.java
@@ -1,9 +1,12 @@
 package com.corona.TestJpa.service.EmployeeServiceImplementation;
 
+
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.corona.TestJpa.dao.EmployeeDao;
@@ -17,8 +20,18 @@ public class EmployeeServiceImpl implements EmployeeService {
 	private EmployeeDao employeeDao;
 
 	@Override
-	public void addEmployee(Employee employee) {
-		employeeDao.save(employee);
+	public ResponseEntity<Boolean> addEmployee(Employee employee) {
+		if(employee==null) {
+			return new ResponseEntity<Boolean>(false,HttpStatus.NO_CONTENT);
+		}
+		try {
+			employeeDao.save(employee);
+			return new ResponseEntity<Boolean>(true,HttpStatus.OK);
+		}
+		catch(Exception e) {
+			return new ResponseEntity<Boolean>(false,HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
 	}
 
 	@Override
@@ -27,14 +40,44 @@ public class EmployeeServiceImpl implements EmployeeService {
 	}
 
 	@Override
-	public void deleteEmployee(long id) {
-		employeeDao.deleteById(id);
+	public ResponseEntity<Boolean> deleteEmployee(long id) {
+		if(id==0) {
+			return new ResponseEntity<Boolean>(false,HttpStatus.NO_CONTENT);
+		}
+		else {
+			Optional<Employee> employee=employeeDao.findById(id);
+			if(employee==null) {
+				return new ResponseEntity<Boolean>(false,HttpStatus.NOT_FOUND);
+			}
+		}
+		try{
+			employeeDao.deleteById(id);
+			return new ResponseEntity<Boolean>(true,HttpStatus.OK);
+		}
+		catch(Exception e) {
+			return new ResponseEntity<Boolean>(false,HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
 	@Override
-	public Optional<Employee> getEmployee(long id) {
-		var temp = employeeDao.findById(id);
-		return temp;
+	public ResponseEntity<Optional<Employee>> getEmployee(long id) {
+		if(id==0) {
+			return new ResponseEntity<Optional<Employee>>(HttpStatus.NO_CONTENT);
+		}
+		else {
+			Optional<Employee> employee=employeeDao.findById(id);
+			if(employee==null) {
+				return new ResponseEntity<Optional<Employee>>(HttpStatus.NOT_FOUND);
+			}
+		}
+		try {
+			Optional<Employee> employee=employeeDao.findById(id);
+			return new ResponseEntity<Optional<Employee>>(employee,HttpStatus.OK);
+		}
+		catch(Exception e) {
+			return new ResponseEntity<Optional<Employee>>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
 	}
 
 }
